refactor: clear program to start over
